+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "axios": "^1.9.0",
     "mongoose": "^8.15.0",
     "next": "15.3.2",
     "react": "^19.1.0",
