+++ src/app/product/api/review/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
-import { pingDatabase } from "@/database";
-import { Review } from "@/database/models/productReview";
+import {connectToDatabase} from "@/database";
+import { ReviewSchema } from "@/database/models/productReview";
 import mongoose from "mongoose";
 
 export async function POST(req: NextRequest){
@@ -13,7 +13,7 @@ export async function POST(req: NextRequest){
             return NextResponse.json({error:"Comment cannot be empty!!!"},{status: 400})
         }
 
-        await pingDatabase()
+        await connectToDatabase();
         // if(!token){
         //     return NextResponse.redirect(new URL("login", req.url))
         // }
@@ -21,7 +21,7 @@ export async function POST(req: NextRequest){
         const productId = new mongoose.Types.ObjectId(`${""}`);   
         const userId = new mongoose.Types.ObjectId( `${""}`);
 
-        const newomment = new Review({productId, userId, comment, rating:stars});
+        const newomment = new ReviewSchema({productId, userId, comment, rating:stars});
         await newomment.save();
 
         return NextResponse.json({message: "Product added successfully"}, {status: 200})
